Rename misnamed 'to_coments' to 'to_comments'
